Restrict development dependencies for Ruby 1.9 and 2.0 compatibility.
Newer versions of nokogiri and tins aren't compatible with both Ruby 1.9
and 2.0 (which, for now, are still our target Ruby versinos). It
shouldn't be necessary to manually restrict the allowed versions like
this, since Bundler shouldn't try to install a version that is
incompatible with the current Ruby, but that functionality doesn't seem
to be working reliably, at least on Travis.